Moved the "Container" component from the hrInfo to the common folder.
